+++ SavGremlin/Source/SavingsAccount.cs
@@ -8,7 +8,7 @@ namespace SavGremlin.Source
     public class SavingsAccount
     {
         public String accountName;
-        public double accountValue;
+        public Decimal accountValue;
 
         public static SavingsAccount inflateFromString(String input)
         {
@@ -17,7 +17,7 @@ namespace SavGremlin.Source
             String[] values = input.Split(':');
 
             result.accountName = values[0];
-            result.accountValue = Convert.ToDouble(values[1]);
+            result.accountValue = Convert.ToDecimal(values[1]);
 
             return result;
         }
